requirements_api_lean.txt
# Docker build - fast, small image (~200MB not 4GB)

# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets>=12.0

# Core pipeline
langgraph>=0.2.0
langchain-anthropic>=0.2.0
anthropic>=0.40.0

# Biology
biopython>=1.84
requests>=2.32.0
pandas>=2.2.0
numpy>=2.0.0

# Data models
pydantic>=2.8.0

# Database
supabase>=2.7.0

# Utils
python-dotenv>=1.0.0

#JWT handling
PyJWT>=2.0.0
python-jose[cryptography]>=3.3.0